Added store earnings statistic and fixed some bugs

+++ src/main/java/ua/com/alevel/controller/ClientCheckController.java
@@ -140,7 +140,7 @@ public class ClientCheckController {
 
                 Token token = Token.create(params);
                 stripeService.charge(
-                        (int) (phoneInstanceService.findPriceForClientCheckId(createOrderParams.getCheckId()) * 100),
+                        phoneInstanceService.findPriceForClientCheckId(createOrderParams.getCheckId()) * 100,
                         "usd",
                         SecurityContextHolder.getContext().getAuthentication().getName(),
                         token.getId()
